requirements.txt
flask~=2.0.1
newspaper3k
nltk~=3.6.2
requests~=2.26.0
beautifulsoup4~=4.12.2
python-dotenv